Fix conosle warnings
