+++ _api/farcaster/casts/channels/bestrecent.ts
@@ -2,7 +2,7 @@ import type { VercelRequest, VercelResponse } from '@vercel/node';
 
 import { InternalServerError } from '../../../../api-lib/HttpError.ts';
 import { fetchCastsForChannel } from '../../../../api-lib/neynar.ts';
-import { getBestCast } from '../../../../api-lib/openai.ts';
+import { getBestCast } from '../../../../api-lib/openai';
 import { describeImage } from '../../../../src/features/ai/replicate.ts';
 
 export default async function handler(req: VercelRequest, res: VercelResponse) {
@@ -36,6 +36,8 @@ export default async function handler(req: VercelRequest, res: VercelResponse) {
 
     const embeds = cast?.embeds;
 
+    const url = `https://warpcast.com/~/conversations/${cast?.hash}`;
+
     let imageDesc = '';
     if (embeds) {
       try {
@@ -54,13 +56,16 @@ export default async function handler(req: VercelRequest, res: VercelResponse) {
           );
 
           imageDesc = await describeImage(imageUrl);
+        } else {
+          // eslint-disable-next-line no-console
+          console.log('No image URL found in embeds', embeds);
         }
       } catch (e) {
         console.error('image decribing failed', e);
       }
     }
 
-    return res.status(200).json({ cast, reasoning, imageDesc });
+    return res.status(200).json({ cast, reasoning, imageDesc, url });
   } catch (e) {
     console.error(e);
     throw new InternalServerError('Error occurred searching profiles', e);
